Remvoe Changes
- remove changes associated with navigation and maps events
- no longer expose more of our sdk
- clean-up
