+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/Event.java
@@ -12,9 +12,9 @@ public abstract class Event implements Parcelable {
     NAV_FEEDBACK, NAV_FASTER_ROUTE
   }
 
-  public static EnumSet<Type> mapGestureEventTypes = EnumSet.of(Type.MAP_CLICK, Type.MAP_DRAGEND);
-  public static EnumSet<Type> navigationEventTypes = EnumSet.of(Type.NAV_DEPART, Type.NAV_ARRIVE, Type.NAV_CANCEL,
+  static EnumSet<Type> mapGestureEventTypes = EnumSet.of(Type.MAP_CLICK, Type.MAP_DRAGEND);
+  static EnumSet<Type> navigationEventTypes = EnumSet.of(Type.NAV_DEPART, Type.NAV_ARRIVE, Type.NAV_CANCEL,
     Type.NAV_REROUTE, Type.NAV_FEEDBACK, Type.NAV_FASTER_ROUTE);
 
-  public abstract Type obtainType();
+  abstract Type obtainType();
 }
