+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/LocationEvent.java
@@ -44,7 +44,7 @@ class LocationEvent extends Event implements Parcelable {
   }
 
   @Override
-  public Type obtainType() {
+  Type obtainType() {
     return Type.LOCATION;
   }
