+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/MapBuildEvent.java
@@ -1,0 +1,6 @@
+package com.mapbox.android.telemetry;
+
+
+interface MapBuildEvent {
+  Event build(MapState mapState);
+}
