+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/MapState.java
@@ -1,4 +1,4 @@
-package com.mapbox.android.telemetry.maps.events;
+package com.mapbox.android.telemetry;
 
 public class MapState {
   private double latitude;
