+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/NavigationArriveEvent.java
@@ -17,7 +17,7 @@ class NavigationArriveEvent extends Event implements Parcelable {
   }
 
   @Override
-  public Type obtainType() {
+  Type obtainType() {
     return Type.NAV_ARRIVE;
   }
