+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/NavigationCancelEvent.java
@@ -20,7 +20,7 @@ class NavigationCancelEvent extends Event implements Parcelable {
   }
 
   @Override
-  public Type obtainType() {
+  Type obtainType() {
     return Type.NAV_CANCEL;
   }
