+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/NavigationFeedbackEvent.java
@@ -28,7 +28,7 @@ class NavigationFeedbackEvent extends Event implements Parcelable {
   }
 
   @Override
-  public Type obtainType() {
+  Type obtainType() {
     return Type.NAV_FEEDBACK;
   }
