+++ libtelemetry/src/main/java/com/mapbox/android/telemetry/NavigationRerouteEvent.java
@@ -31,7 +31,7 @@ class NavigationRerouteEvent extends Event implements Parcelable {
   }
 
   @Override
-  public Type obtainType() {
+  Type obtainType() {
     return Type.NAV_REROUTE;
   }
