--- libtelemetry/src/main/java/com/mapbox/android/telemetry/maps/events/MapBuildEvent.java
@@ -1,8 +1,0 @@
-package com.mapbox.android.telemetry.maps.events;
-
-
-import com.mapbox.android.telemetry.Event;
-
-interface MapBuildEvent {
-  Event build(MapState mapState);
-}
